pyproject.toml
[project]
name = "gemeni-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "assemblyai>=0.41.5",
    "fastapi>=0.115.13",
    "google-cloud-texttospeech>=2.27.0",
    "google-genai>=1.21.1",
    "langchain-google-genai>=2.1.5",
    "pillow>=11.2.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
    "youtube-transcript-api>=1.1.0",
    "yt-dlp>=2025.6.9",
]
